# modulo_promocional_api/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Modulos-Promocionales-api
  description: Microservicio de gestion de peticiones de Informacion e inyeccion de
    Datos
  version: 1.0.11
servers:
- url: /rest/modulos-promocionales-api/v1.0/
...
paths:
...
  /get/finance/buro:
    post:
      tags:
      - Financiero
      description: Consulta de Buro
      operationId: get_buro
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetBuro'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetBuro_data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.financiero_controller
  /get/finance/modos-pago:
    post:
      tags:
      - Financiero
      description: Consulta de Modos de Pago
      operationId: get_modospago
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetFormasPago'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetFormasPago_data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.financiero_controller
  /get/finance/dias-gozados:
    post:
      tags:
      - Financiero
      description: Consulta de Dias Gozados
      operationId: get_diasgozados
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetDiasGozados'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetDiasGozados_data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.financiero_controller
  /get/finance/upgrade:
    post:
      tags:
      - Financiero
      description: Consulta de Datos Upgrade
      operationId: get_upgrade
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetUpgrade'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetUpgrade_data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.financiero_controller
  /get/finance/precio-regular:
    post:
      tags:
      - Financiero
      description: Consulta de Precios de Productos
      operationId: get_precioregular
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetPrecioRegular'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetPrecioRegular_data'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseError'
      x-openapi-router-controller: swagger_server.controllers.financiero_controller
...
components:
  schemas:
...
    RequestGetBuro:
      required:
      - channel
      - externalTransactionId
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ALL_
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    RequestGetFormasPago:
      required:
      - channel
      - externalTransactionId
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ALL_
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    RequestGetDiasGozados:
      required:
      - channel
      - externalTransactionId
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ALL_
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    RequestGetPrecioRegular:
      required:
      - TARIFFPLANVARIANT
      - channel
      - externalTransactionId
      - id_Prod
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ' ALL_'
        TARIFFPLANVARIANT:
          type: integer
          example: 128213
        id_Prod:
          type: integer
          example: 128213
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    RequestGetUpgrade:
      required:
      - TARIFFPLAN
      - TARIFFPLANVARIANT
      - channel
      - externalTransactionId
      - type
      properties:
        channel:
          type: string
          example: modulos-promocionales-web
        type:
          type: string
          example: ALL_
        TARIFFPLAN:
          type: integer
          example: 8909
        TARIFFPLANVARIANT:
          type: integer
          example: 128213
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
...
    ResponseGetBuro_data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        BURO:
          $ref: '#/components/schemas/ResponseGetBuro_data_BURO'
      example:
        errorCode: 0
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
        BURO:
          ID: 0
          NAME: NAME
    ResponseGetFormasPago_data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        MPAGOS:
          $ref: '#/components/schemas/ResponseGetBuro_data_BURO'
      example:
        MPAGOS:
          ID: 0
          NAME: NAME
        errorCode: 0
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
    ResponseGetDiasGozados_data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        DIAS_GOZADOS:
          $ref: '#/components/schemas/ResponseGetBuro_data_BURO'
      example:
        errorCode: 0
        DIAS_GOZADOS:
          ID: 0
          NAME: NAME
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
    ResponseGetPrecioRegular_data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        PRECIO_REGULAR:
          $ref: '#/components/schemas/ResponseGetPrecioRegular_data_PRECIO_REGULAR'
      example:
        PRECIO_REGULAR:
          PRECIO: 0.8008282
        errorCode: 0
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
    ResponseGetUpgrade_data:
      type: object
      properties:
        errorCode:
          type: integer
          example: 0
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: OK
        UPGRADE:
          $ref: '#/components/schemas/ResponseGetUpgrade_data_UPGRADE'
      example:
        UPGRADE:
          PLAN_UPGRADE: PLAN_UPGRADE
          ID: 0
        errorCode: 0
        internalTransactionId: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId: 82fb166d-c273-4840-9008-c895cf9de255
        message: OK
...
    ResponseError:
      type: object
      properties:
        errorCode:
          type: integer
          example: -1
        internalTransactionId:
          type: string
          example: 03c33780-88cf-46cf-ba22-925e4d5be471
        externalTransactionId:
          type: string
          example: 82fb166d-c273-4840-9008-c895cf9de255
        message:
          type: string
          example: mensaje de error
...
    ResponseGetBuro_data_BURO:
      type: object
      properties:
        ID:
          type: integer
        NAME:
          type: string
      example:
        ID: 0
        NAME: NAME
    ResponseGetPrecioRegular_data_PRECIO_REGULAR:
      type: object
      properties:
        PRECIO:
          type: number
          format: float
      example:
        PRECIO: 0.8008282
    ResponseGetUpgrade_data_UPGRADE:
      type: object
      properties:
        ID:
          type: integer
        PLAN_UPGRADE:
          type: string
      example:
        PLAN_UPGRADE: PLAN_UPGRADE
        ID: 0